XHR Slides from the remote server.
This is currently looking in presentation/ for an slides.json, which is an
array of file names to be loaded via XHR from that same directory. When they
are all XHR'd, they are mapped in to the viewTypes array and the Pane is reset
to the first page.

This should mostly work, but it's doing things like eval'ing and functions
within those objects I'm creating aren't working for some reason. :\